+++ ui/src/components/Loading/Loading.tsx
@@ -1,4 +1,4 @@
-import "../styles/loading.css";
+import "../../styles/loading.css";
 
 export default function Loading() {
     return (
